fix: коректировка модулей
